+++ client/src/App.js
@@ -3,13 +3,13 @@ import React from 'react'
 import './App.css'
 import SignIn from './components/SignIn/SignIn'
 import ProtectedRoute from './components/ProtectedRoute/ProtectedRoute'
-import Register from './components/Register/Register'
+import SignUp from './components/SignUp/SignUp'
 
 export default function App () {
   return (
     <div>
       <Switch>
-        <Route exact path='/sign-up' component={Register}/>
+        <Route exact path='/sign-up' component={SignUp}/>
         <Route exact path='/sign-in' component={SignIn}/>
         <Route exact path='/*' component={ProtectedRoute}/>
       </Switch>
