--- client/src/components/Register/Register.js
@@ -1,223 +1,0 @@
-import React, {useState} from 'react'
-import Avatar from '@material-ui/core/Avatar'
-import Button from '@material-ui/core/Button'
-import CssBaseline from '@material-ui/core/CssBaseline'
-import TextField from '@material-ui/core/TextField'
-import FormControlLabel from '@material-ui/core/FormControlLabel'
-import Checkbox from '@material-ui/core/Checkbox'
-import Link from '@material-ui/core/Link'
-import Grid from '@material-ui/core/Grid'
-import Box from '@material-ui/core/Box'
-import LockOutlinedIcon from '@material-ui/icons/LockOutlined'
-import Typography from '@material-ui/core/Typography'
-import {makeStyles} from '@material-ui/core/styles'
-import Container from '@material-ui/core/Container'
-import GenderSelect from '../GenderSelect/GenderSelect'
-import {NavLink} from 'react-router-dom'
-import {useDispatch, useSelector} from 'react-redux'
-import {createUser} from '../../actions/profileActions'
-import Redirect from 'react-router-dom/es/Redirect'
-
-function Copyright () {
-  return (
-    <Typography variant="body2" color="textSecondary" align="center">
-      {'Copyright © '}
-      <Link color="inherit" href="https://material-ui.com/">
-                Your Website
-      </Link>{' '}
-      {new Date().getFullYear()}
-      {'.'}
-    </Typography>
-  )
-}
-
-const useStyles = makeStyles(theme => ({
-  paper: {
-    marginTop: theme.spacing(8),
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center'
-  },
-  backdrop: {
-    zIndex: theme.zIndex.drawer + 1,
-    color: '#fff'
-  },
-  selectContainer: {
-    display: 'flex'
-
-  },
-  avatar: {
-    margin: theme.spacing(1),
-    backgroundColor: theme.palette.secondary.main
-  },
-  form: {
-    width: '100%', // Fix IE 11 issue.
-    marginTop: theme.spacing(3)
-  },
-  submit: {
-    margin: theme.spacing(3, 0, 2)
-  }
-}))
-
-export default function SignUp () {
-  const classes = useStyles()
-
-  const [fName, setfName] = useState('')
-  const [lName, setlName] = useState('')
-  const [password, setPassword] = useState('')
-  const [email, setEmail] = useState('')
-  const [username, setUsername] = useState('')
-  const [date, setDate] = useState('')
-  const [gender, setGender] = React.useState('')
-
-  const dispatch = useDispatch()
-  const currentUser = useSelector(state => state.users.currentUser)
-  const submitValue = event => {
-    event.preventDefault()
-    const frmdetails = {
-      'firstName': fName,
-      'lastName': lName,
-      'email': email,
-      'password': password,
-      'username': username,
-      'gender': gender,
-      'birthDate': date
-    }
-    const formData = new FormData()
-    formData.append('username', username)
-    formData.append('password', password)
-
-    dispatch(createUser(frmdetails, formData))
-  }
-  if (currentUser) {
-    return (
-      <Redirect to="/profile"/>
-    )
-  }
-  return (
-    <Container component="main" maxWidth="xs">
-      <CssBaseline/>
-      <div className={classes.paper}>
-        <Avatar className={classes.avatar}>
-          <LockOutlinedIcon/>
-        </Avatar>
-        <Typography component="h1" variant="h5">
-                    Sign up
-        </Typography>
-        <form className={classes.form} noValidate>
-          <Grid container spacing={2}>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                onChange={e => setfName(e.target.value)}
-                autoComplete="fname"
-                name="firstName"
-                variant="outlined"
-                required
-                fullWidth
-                id="firstName"
-                label="First Name"
-                autoFocus
-              />
-            </Grid>
-            <Grid item xs={12} sm={6}>
-              <TextField
-                onChange={e => setlName(e.target.value)}
-                variant="outlined"
-                required
-                fullWidth
-                id="lastName"
-                label="Last Name"
-                name="lastName"
-
-                autoComplete="lname"
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <TextField
-                onChange={e => setUsername(e.target.value)}
-                variant="outlined"
-                required
-                fullWidth
-                name="username"
-                label="Username"
-                id="username"
-                autoComplete="current-password"
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <TextField
-                onChange={e => setEmail(e.target.value)}
-                variant="outlined"
-                required
-                fullWidth
-                id="email"
-                label="Email Address"
-                name="email"
-                autoComplete="email"
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <TextField
-                onChange={e => setPassword(e.target.value)}
-                variant="outlined"
-                required
-                fullWidth
-                name="password"
-                label="Password"
-                type="password"
-                id="password"
-                autoComplete="current-password"
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <div className={classes.selectContainer}>
-                <form className={classes.container} noValidate>
-                  <TextField
-                    onChange={e => setDate(e.target.value)}
-                    id="date"
-                    label="Birthday"
-                    type="date"
-                    defaultValue=""
-                    className={classes.textField}
-                    InputLabelProps={{
-                      shrink: true
-                    }}
-                  />
-                </form>
-                <GenderSelect setGender={setGender}/>
-              </div>
-
-            </Grid>
-            <Grid item xs={12}>
-              <FormControlLabel
-                control={<Checkbox value="allowExtraEmails" color="primary"/>}
-                label="I want to receive inspiration, marketing promotions and updates via email."
-              />
-            </Grid>
-          </Grid>
-          <Button
-            type="submit"
-            fullWidth
-            variant="contained"
-            color="primary"
-            className={classes.submit}
-            onClick={submitValue}
-          >
-                        Sign Up
-          </Button>
-
-          <Grid container justify="flex-end">
-            <Grid item>
-              <NavLink to="/sign-in" href="#" variant="body2">
-                                Already have an account? Sign in
-              </NavLink>
-            </Grid>
-          </Grid>
-        </form>
-      </div>
-      <Box mt={5}>
-        <Copyright/>
-      </Box>
-    </Container>
-  )
-}
