+++ client/src/components/SignIn/SignIn.js
@@ -7,12 +7,10 @@ import FormControlLabel from '@material-ui/core/FormControlLabel'
 import Checkbox from '@material-ui/core/Checkbox'
 import Link from '@material-ui/core/Link'
 import Grid from '@material-ui/core/Grid'
-import Box from '@material-ui/core/Box'
 import LockOutlinedIcon from '@material-ui/icons/LockOutlined'
 import Typography from '@material-ui/core/Typography'
 import { makeStyles } from '@material-ui/core/styles'
 import Container from '@material-ui/core/Container'
-import Copyright from '../Copyright/copyright'
 import {NavLink} from 'react-router-dom'
 import {logUser} from '../../actions/profileActions'
 import {useDispatch, useSelector} from 'react-redux'
@@ -74,11 +72,10 @@ export default function SignIn () {
             margin='normal'
             required
             fullWidth
-            id='email'
-            label='Email Address'
-            name='email'
+            id='username'
+            label='Username'
+            name='username'
             onChange={e => setUsername(e.target.value)}
-            autoComplete='email'
             autoFocus
           />
           <TextField
@@ -121,9 +118,6 @@ export default function SignIn () {
           </Grid>
         </form>
       </div>
-      <Box mt={8}>
-        <Copyright />
-      </Box>
     </Container>
   )
 }
